requirements.txt
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.5
django-extra-fields==1.0.0
django-filter==1.1.0
django-multiselectfield==0.1.8
django-rest-swagger==2.2.0
djangorestframework==3.8.2
feedparser==5.2.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.1.0
pycparser==2.19
pyotp==2.2.6
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
